+++ src/brasil/gov/portal/patches.py
@@ -4,7 +4,10 @@ from brasil.gov.portal.logger import logger
 from plone.app.contenttypes.content import Link
 from plone.i18n.normalizer.interfaces import IIDNormalizer
 from plone.outputfilters.filters import resolveuid_and_caption as base
+from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone import utils
+from Products.CMFPlone.browser.navigation import get_id
+from Products.CMFPlone.browser.navigation import get_view_url
 from zope.component import getMultiAdapter
 from zope.component import queryUtility
 
@@ -129,6 +132,92 @@ def decoratorFactory(self, node):
     return newNode
 
     logger.info('Patched Products.CMFPlone.browser.navtree.SitemapNavtreeStrategy:decoratorFactory:191')
+
+
+def topLevelTabs(self, actions=None, category='portal_tabs'):
+    """
+    Esse método veio de
+
+        https://github.com/plone/Products.CMFPlone/blob/4.3.18/Products/CMFPlone/browser/navigation.py
+
+    Com o objetivo de alterar o comportamento de get_link_url:
+
+        https://github.com/plone/Products.CMFPlone/blob/4.3.18/Products/CMFPlone/browser/navigation.py#L195
+
+    É a única customização no método topLevelTabs.
+
+    Em tese, poderia ter sido feito um patch apenas no método get_link_url, mas
+    como o formato é uma função de uma função e a documentação de como isso deve
+    ser feito consideramos complexa:
+
+        https://stackoverflow.com/questions/27550228/can-you-patch-just-a-nested-function-with-closure-or-must-the-whole-outer-fun/27550237
+
+    Decidimos ir pelo caminho de copiar a topLevelTabs.
+    """
+    context = aq_inner(self.context)
+
+    mtool = getToolByName(context, 'portal_membership')
+    member = mtool.getAuthenticatedMember().id
+
+    portal_properties = getToolByName(context, 'portal_properties')
+    self.navtree_properties = getattr(portal_properties,
+                                      'navtree_properties')
+    self.site_properties = getattr(portal_properties,
+                                   'site_properties')
+    self.portal_catalog = getToolByName(context, 'portal_catalog')
+
+    if actions is None:
+        context_state = getMultiAdapter((context, self.request),
+                                        name=u'plone_context_state')
+        actions = context_state.actions(category)
+
+    # Build result dict
+    result = []
+    # first the actions
+    if actions is not None:
+        for actionInfo in actions:
+            data = actionInfo.copy()
+            data['name'] = data['title']
+            result.append(data)
+
+    # check whether we only want actions
+    if self.site_properties.getProperty('disable_folder_sections', False):
+        return result
+
+    query = self._getNavQuery()
+
+    rawresult = self.portal_catalog.searchResults(query)
+
+    def get_link_url(item):
+        linkremote = item.getRemoteUrl and not member == item.Creator  # NOQA
+        if linkremote:
+            # patch: Substitui o path pela url do site.
+            remote_url_utils = getMultiAdapter(
+                (item, item.REQUEST),
+                name=u'remote_url_utils',
+            )
+            remote_url = item.getRemoteUrl and item.getRemoteUrl or None
+            remote_url_transform = remote_url_utils.remote_url_transform(
+                item.getPath(),
+                remote_url,
+            )
+            return (get_id(item), remote_url_transform)
+            # patch
+        else:
+            return False
+
+    # now add the content to results
+    idsNotToList = self.navtree_properties.getProperty('idsNotToList', ())
+    for item in rawresult:
+        if not (item.getId in idsNotToList or item.exclude_from_nav):
+            id, item_url = get_link_url(item) or get_view_url(item)
+            data = {'name': utils.pretty_title_or_id(context, item),
+                    'id': item.getId,
+                    'url': item_url,
+                    'description': item.Description}
+            result.append(data)
+
+    return result
 
 
 def run():